requirements.txt
streamlit==1.28.2
torch>=2.0.0
opencv-python
pillow<11.0.0
numpy
matplotlib
pyyaml
